rest/requirements.txt
coreapi==2.1.1
Django==1.10.3
django-filter==0.15.3
djangorestframework==3.5.2
gunicorn==19.6.0
itypes==1.1.0
Markdown==2.6.7
MySQL-python==1.2.5
requests==2.12.5
uritemplate==3.0.0